+++ test-common/test-common-api/src/main/java/org/apache/kafka/common/test/api/Type.java
@@ -16,6 +16,7 @@
  */
 package org.apache.kafka.common.test.api;
 
+import org.apache.kafka.common.test.RaftClusterInvocationContext;
 import org.junit.jupiter.api.extension.TestTemplateInvocationContext;
